% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_choice_rule.R
\name{BaseRule}
\alias{BaseRule}
\title{BaseRule}
\description{
Abstract class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-to_list}{\code{BaseRule$to_list()}}
\item \href{#method-format}{\code{BaseRule$format()}}
\item \href{#method-clone}{\code{BaseRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert class to list ready to be translated for
             Amazon States Language \url{https://states-language.net/spec.html}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRule$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
