% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_states.R
\name{Catch}
\alias{Catch}
\title{Catch base Class}
\description{
A class for creating a Catch block.
}
\section{Super class}{
\code{\link[stepfunctions:Block]{stepfunctions::Block}} -> \code{Catch}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Catch$new()}}
\item \href{#method-allowed_fields}{\code{Catch$allowed_fields()}}
\item \href{#method-to_list}{\code{Catch$to_list()}}
\item \href{#method-clone}{\code{Catch$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="format">}\href{../../stepfunctions/html/Block.html#method-format}{\code{stepfunctions::Block$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="is_field_allowed">}\href{../../stepfunctions/html/Block.html#method-is_field_allowed}{\code{stepfunctions::Block$is_field_allowed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="to_json">}\href{../../stepfunctions/html/Block.html#method-to_json}{\code{stepfunctions::Block$to_json()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Catch block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catch$new(next_step, error_equals, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{next_step}}{(State or Chain): Next state or chain to transition to.}

\item{\code{error_equals}}{(list(str)): Non-empty list of strings, which match
`Error Names <https://states-language.net/spec.html#error-names>`_.
When a state reports an error, the interpreter scans through the catchers
and, when the Error Name appears in the value of of a catcher's `error_equals`
field, transitions to the `next_step` described in the catcher.}

\item{\code{...}}{: Extra field names to pass to Block class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allowed_fields"></a>}}
\if{latex}{\out{\hypertarget{method-allowed_fields}{}}}
\subsection{Method \code{allowed_fields()}}{
allowed extra fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catch$allowed_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert class to list ready to be translated for
             `Amazon States Language <https://states-language.net/spec.html>`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catch$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
