% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_choice_rule.R
\name{ChoiceRule}
\alias{ChoiceRule}
\title{ChoiceRule class}
\description{
Factory class for creating a choice rule.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StringEquals}{\code{ChoiceRule$StringEquals()}}
\item \href{#method-StringEqualsPath}{\code{ChoiceRule$StringEqualsPath()}}
\item \href{#method-StringLessThan}{\code{ChoiceRule$StringLessThan()}}
\item \href{#method-StringLessThanPath}{\code{ChoiceRule$StringLessThanPath()}}
\item \href{#method-StringGreaterThan}{\code{ChoiceRule$StringGreaterThan()}}
\item \href{#method-StringGreaterThanPath}{\code{ChoiceRule$StringGreaterThanPath()}}
\item \href{#method-StringLessThanEquals}{\code{ChoiceRule$StringLessThanEquals()}}
\item \href{#method-StringLessThanEqualsPath}{\code{ChoiceRule$StringLessThanEqualsPath()}}
\item \href{#method-StringGreaterThanEquals}{\code{ChoiceRule$StringGreaterThanEquals()}}
\item \href{#method-StringGreaterThanEqualsPath}{\code{ChoiceRule$StringGreaterThanEqualsPath()}}
\item \href{#method-NumericEquals}{\code{ChoiceRule$NumericEquals()}}
\item \href{#method-NumericEqualsPath}{\code{ChoiceRule$NumericEqualsPath()}}
\item \href{#method-NumericLessThan}{\code{ChoiceRule$NumericLessThan()}}
\item \href{#method-NumericLessThanPath}{\code{ChoiceRule$NumericLessThanPath()}}
\item \href{#method-NumericGreaterThan}{\code{ChoiceRule$NumericGreaterThan()}}
\item \href{#method-NumericGreaterThanPath}{\code{ChoiceRule$NumericGreaterThanPath()}}
\item \href{#method-NumericLessThanEquals}{\code{ChoiceRule$NumericLessThanEquals()}}
\item \href{#method-NumericLessThanEqualsPath}{\code{ChoiceRule$NumericLessThanEqualsPath()}}
\item \href{#method-NumericGreaterThanEquals}{\code{ChoiceRule$NumericGreaterThanEquals()}}
\item \href{#method-NumericGreaterThanEqualsPath}{\code{ChoiceRule$NumericGreaterThanEqualsPath()}}
\item \href{#method-BooleanEquals}{\code{ChoiceRule$BooleanEquals()}}
\item \href{#method-BooleanEqualsPath}{\code{ChoiceRule$BooleanEqualsPath()}}
\item \href{#method-TimestampEquals}{\code{ChoiceRule$TimestampEquals()}}
\item \href{#method-TimestampEqualsPath}{\code{ChoiceRule$TimestampEqualsPath()}}
\item \href{#method-TimestampLessThan}{\code{ChoiceRule$TimestampLessThan()}}
\item \href{#method-TimestampLessThanPath}{\code{ChoiceRule$TimestampLessThanPath()}}
\item \href{#method-TimestampGreaterThan}{\code{ChoiceRule$TimestampGreaterThan()}}
\item \href{#method-TimestampGreaterThanPath}{\code{ChoiceRule$TimestampGreaterThanPath()}}
\item \href{#method-TimestampLessThanEquals}{\code{ChoiceRule$TimestampLessThanEquals()}}
\item \href{#method-TimestampLessThanEqualsPath}{\code{ChoiceRule$TimestampLessThanEqualsPath()}}
\item \href{#method-TimestampGreaterThanEquals}{\code{ChoiceRule$TimestampGreaterThanEquals()}}
\item \href{#method-TimestampGreaterThanEqualsPath}{\code{ChoiceRule$TimestampGreaterThanEqualsPath()}}
\item \href{#method-IsNull}{\code{ChoiceRule$IsNull()}}
\item \href{#method-IsPresent}{\code{ChoiceRule$IsPresent()}}
\item \href{#method-IsString}{\code{ChoiceRule$IsString()}}
\item \href{#method-IsNumeric}{\code{ChoiceRule$IsNumeric()}}
\item \href{#method-IsTimestamp}{\code{ChoiceRule$IsTimestamp()}}
\item \href{#method-IsBoolean}{\code{ChoiceRule$IsBoolean()}}
\item \href{#method-StringMatches}{\code{ChoiceRule$StringMatches()}}
\item \href{#method-And}{\code{ChoiceRule$And()}}
\item \href{#method-Or}{\code{ChoiceRule$Or()}}
\item \href{#method-Not}{\code{ChoiceRule$Not()}}
\item \href{#method-clone}{\code{ChoiceRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringEquals"></a>}}
\if{latex}{\out{\hypertarget{method-StringEquals}{}}}
\subsection{Method \code{StringEquals()}}{
Creates a rule with the `StringEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-StringEqualsPath}{}}}
\subsection{Method \code{StringEqualsPath()}}{
Creates a rule with the `StringEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringLessThan"></a>}}
\if{latex}{\out{\hypertarget{method-StringLessThan}{}}}
\subsection{Method \code{StringLessThan()}}{
Creates a rule with the `StringLessThan` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringLessThan(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringLessThan` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringLessThanPath"></a>}}
\if{latex}{\out{\hypertarget{method-StringLessThanPath}{}}}
\subsection{Method \code{StringLessThanPath()}}{
Creates a rule with the `StringLessThanPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringLessThanPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringLessThanPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringGreaterThan"></a>}}
\if{latex}{\out{\hypertarget{method-StringGreaterThan}{}}}
\subsection{Method \code{StringGreaterThan()}}{
Creates a rule with the `StringGreaterThan` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringGreaterThan(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringGreaterThan` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringGreaterThanPath"></a>}}
\if{latex}{\out{\hypertarget{method-StringGreaterThanPath}{}}}
\subsection{Method \code{StringGreaterThanPath()}}{
Creates a rule with the `StringGreaterThanPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringGreaterThanPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringGreaterThanPath` operator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringLessThanEquals"></a>}}
\if{latex}{\out{\hypertarget{method-StringLessThanEquals}{}}}
\subsection{Method \code{StringLessThanEquals()}}{
Creates a rule with the `StringLessThanEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringLessThanEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringLessThanEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringLessThanEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-StringLessThanEqualsPath}{}}}
\subsection{Method \code{StringLessThanEqualsPath()}}{
Creates a rule with the `StringLessThanEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringLessThanEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringLessThanEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringGreaterThanEquals"></a>}}
\if{latex}{\out{\hypertarget{method-StringGreaterThanEquals}{}}}
\subsection{Method \code{StringGreaterThanEquals()}}{
Creates a rule with the `StringGreaterThanEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringGreaterThanEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringGreaterThanEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringGreaterThanEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-StringGreaterThanEqualsPath}{}}}
\subsection{Method \code{StringGreaterThanEqualsPath()}}{
Creates a rule with the `StringGreaterThanEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringGreaterThanEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringGreaterThanEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericEquals"></a>}}
\if{latex}{\out{\hypertarget{method-NumericEquals}{}}}
\subsection{Method \code{NumericEquals()}}{
Creates a rule with the `NumericEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(int): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-NumericEqualsPath}{}}}
\subsection{Method \code{NumericEqualsPath()}}{
Creates a rule with the `NumericEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericLessThan"></a>}}
\if{latex}{\out{\hypertarget{method-NumericLessThan}{}}}
\subsection{Method \code{NumericLessThan()}}{
Creates a rule with the `NumericLessThan` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericLessThan(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(int): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericLessThan` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericLessThanPath"></a>}}
\if{latex}{\out{\hypertarget{method-NumericLessThanPath}{}}}
\subsection{Method \code{NumericLessThanPath()}}{
Creates a rule with the `NumericLessThanPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericLessThanPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericLessThanPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericGreaterThan"></a>}}
\if{latex}{\out{\hypertarget{method-NumericGreaterThan}{}}}
\subsection{Method \code{NumericGreaterThan()}}{
Creates a rule with the `NumericGreaterThan` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericGreaterThan(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(int): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericGreaterThan` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericGreaterThanPath"></a>}}
\if{latex}{\out{\hypertarget{method-NumericGreaterThanPath}{}}}
\subsection{Method \code{NumericGreaterThanPath()}}{
Creates a rule with the `NumericGreaterThanPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericGreaterThanPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericGreaterThanPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericLessThanEquals"></a>}}
\if{latex}{\out{\hypertarget{method-NumericLessThanEquals}{}}}
\subsection{Method \code{NumericLessThanEquals()}}{
Creates a rule with the `NumericLessThanEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericLessThanEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(int): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericLessThanEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericLessThanEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-NumericLessThanEqualsPath}{}}}
\subsection{Method \code{NumericLessThanEqualsPath()}}{
Creates a rule with the `NumericLessThanEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericLessThanEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericLessThanEqualsPath` operator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericGreaterThanEquals"></a>}}
\if{latex}{\out{\hypertarget{method-NumericGreaterThanEquals}{}}}
\subsection{Method \code{NumericGreaterThanEquals()}}{
Creates a rule with the `NumericGreaterThanEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericGreaterThanEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(int): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericGreaterThanEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericGreaterThanEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-NumericGreaterThanEqualsPath}{}}}
\subsection{Method \code{NumericGreaterThanEqualsPath()}}{
Creates a rule with the `NumericGreaterThanEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$NumericGreaterThanEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `NumericGreaterThanEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BooleanEquals"></a>}}
\if{latex}{\out{\hypertarget{method-BooleanEquals}{}}}
\subsection{Method \code{BooleanEquals()}}{
Creates a rule with the `BooleanEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$BooleanEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(bool): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `BooleanEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BooleanEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-BooleanEqualsPath}{}}}
\subsection{Method \code{BooleanEqualsPath()}}{
Creates a rule with the `BooleanEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$BooleanEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `BooleanEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampEquals"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampEquals}{}}}
\subsection{Method \code{TimestampEquals()}}{
Creates a rule with the `TimestampEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampEqualsPath}{}}}
\subsection{Method \code{TimestampEqualsPath()}}{
Creates a rule with the `TimestampEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampLessThan"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampLessThan}{}}}
\subsection{Method \code{TimestampLessThan()}}{
Creates a rule with the `TimestampLessThan` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampLessThan(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampLessThan` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampLessThanPath"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampLessThanPath}{}}}
\subsection{Method \code{TimestampLessThanPath()}}{
Creates a rule with the `TimestampLessThanPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampLessThanPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampLessThanPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampGreaterThan"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampGreaterThan}{}}}
\subsection{Method \code{TimestampGreaterThan()}}{
Creates a rule with the `TimestampGreaterThan` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampGreaterThan(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampGreaterThan` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampGreaterThanPath"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampGreaterThanPath}{}}}
\subsection{Method \code{TimestampGreaterThanPath()}}{
Creates a rule with the `TimestampGreaterThanPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampGreaterThanPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampGreaterThanPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampLessThanEquals"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampLessThanEquals}{}}}
\subsection{Method \code{TimestampLessThanEquals()}}{
Creates a rule with the `TimestampLessThanEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampLessThanEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampLessThanEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampLessThanEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampLessThanEqualsPath}{}}}
\subsection{Method \code{TimestampLessThanEqualsPath()}}{
Creates a rule with the `TimestampLessThanEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampLessThanEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampLessThanEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampGreaterThanEquals"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampGreaterThanEquals}{}}}
\subsection{Method \code{TimestampGreaterThanEquals()}}{
Creates a rule with the `TimestampGreaterThanEquals` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampGreaterThanEquals(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Constant value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampGreaterThanEquals` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimestampGreaterThanEqualsPath"></a>}}
\if{latex}{\out{\hypertarget{method-TimestampGreaterThanEqualsPath}{}}}
\subsection{Method \code{TimestampGreaterThanEqualsPath()}}{
Creates a rule with the `TimestampGreaterThanEqualsPath` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$TimestampGreaterThanEqualsPath(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): Path to the value to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `TimestampGreaterThanEqualsPath` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsNull"></a>}}
\if{latex}{\out{\hypertarget{method-IsNull}{}}}
\subsection{Method \code{IsNull()}}{
Creates a rule with the `IsNull` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$IsNull(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(bool): Whether the value at `variable` is equal to the JSON literal null or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `IsNull` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsPresent"></a>}}
\if{latex}{\out{\hypertarget{method-IsPresent}{}}}
\subsection{Method \code{IsPresent()}}{
Creates a rule with the `IsPresent` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$IsPresent(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(bool): Whether a field at `variable` exists in the input or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `IsPresent` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsString"></a>}}
\if{latex}{\out{\hypertarget{method-IsString}{}}}
\subsection{Method \code{IsString()}}{
Creates a rule with the `IsString` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$IsString(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(bool): Whether the value at `variable` is a string or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `IsString` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsNumeric"></a>}}
\if{latex}{\out{\hypertarget{method-IsNumeric}{}}}
\subsection{Method \code{IsNumeric()}}{
Creates a rule with the `IsNumeric` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$IsNumeric(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(bool): Whether the value at `variable` is a number or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `IsNumeric` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsTimestamp"></a>}}
\if{latex}{\out{\hypertarget{method-IsTimestamp}{}}}
\subsection{Method \code{IsTimestamp()}}{
Creates a rule with the `IsTimestamp` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$IsTimestamp(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(bool): Whether the value at `variable` is a timestamp or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `IsTimestamp` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsBoolean"></a>}}
\if{latex}{\out{\hypertarget{method-IsBoolean}{}}}
\subsection{Method \code{IsBoolean()}}{
Creates a rule with the `IsBoolean` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$IsBoolean(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(bool): Whether the value at `variable` is a boolean or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `IsBoolean` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StringMatches"></a>}}
\if{latex}{\out{\hypertarget{method-StringMatches}{}}}
\subsection{Method \code{StringMatches()}}{
Creates a rule with the `StringMatches` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$StringMatches(variable, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{value}}{(str): A string pattern that may contain one or more `*` characters
to compare the value at `variable` to.
The `*` character can be escaped using two backslashes.
The comparison yields true if the variable matches the pattern, where
`*` is a wildcard that matches zero or more characters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rule: Rule with `StringMatches` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-And"></a>}}
\if{latex}{\out{\hypertarget{method-And}{}}}
\subsection{Method \code{And()}}{
Creates a compound rule with the `And` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$And(rules)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rules}}{(list(BaseRule)): List of rules to compound together.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CompoundRule: Compound rule with `And` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Or"></a>}}
\if{latex}{\out{\hypertarget{method-Or}{}}}
\subsection{Method \code{Or()}}{
Creates a compound rule with the `Or` operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$Or(rules)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rules}}{(list(BaseRule)): List of rules to compound together.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CompoundRule: Compound rule with `Or` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Not"></a>}}
\if{latex}{\out{\hypertarget{method-Not}{}}}
\subsection{Method \code{Not()}}{
Creates a negation for a rule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$Not(rule)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rule}}{(BaseRule): Rule to Negate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NotRule: Rule with `Not` operator.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoiceRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
