% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_placeholders.R
\name{ExecutionInput}
\alias{ExecutionInput}
\title{ExecutionInput Class}
\description{
Top-level class for execution input placeholders.
}
\section{Super class}{
\code{\link[stepfunctions:Placeholder]{stepfunctions::Placeholder}} -> \code{ExecutionInput}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExecutionInput$new()}}
\item \href{#method-clone}{\code{ExecutionInput$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Placeholder" data-id="contains">}\href{../../stepfunctions/html/Placeholder.html#method-contains}{\code{stepfunctions::Placeholder$contains()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Placeholder" data-id="get">}\href{../../stepfunctions/html/Placeholder.html#method-get}{\code{stepfunctions::Placeholder$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Placeholder" data-id="get_schema_as_json">}\href{../../stepfunctions/html/Placeholder.html#method-get_schema_as_json}{\code{stepfunctions::Placeholder$get_schema_as_json()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Placeholder" data-id="get_schema_as_list">}\href{../../stepfunctions/html/Placeholder.html#method-get_schema_as_list}{\code{stepfunctions::Placeholder$get_schema_as_list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Placeholder" data-id="to_jsonpath">}\href{../../stepfunctions/html/Placeholder.html#method-to_jsonpath}{\code{stepfunctions::Placeholder$to_jsonpath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Placeholder" data-id="validate">}\href{../../stepfunctions/html/Placeholder.html#method-validate}{\code{stepfunctions::Placeholder$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Placeholder class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionInput$new(schema = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{schema}}{(dict, optional): Schema for the placeholder collection. (default: None)}

\item{\code{...}}{: other parameters from Placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
