% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_fields.R
\docType{data}
\name{Field}
\alias{Field}
\title{AWS Stepfunction state fields}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 33.
}
\usage{
Field
}
\description{
List of Fields to pass to Stepfunction Steps
}
\keyword{datasets}
