% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_states.R
\name{Map}
\alias{Map}
\title{Map state provides the ability to dynamically iterate over a state/subgraph for each entry in a list.}
\description{
A Map state can accept an input with a list of items,
             execute a state or chain for each item in the list, and return
             a list, with all corresponding results of each execution, as its output.
}
\section{Super classes}{
\code{\link[stepfunctions:Block]{stepfunctions::Block}} -> \code{\link[stepfunctions:State]{stepfunctions::State}} -> \code{Map}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Map$new()}}
\item \href{#method-attach_iterator}{\code{Map$attach_iterator()}}
\item \href{#method-allowed_fields}{\code{Map$allowed_fields()}}
\item \href{#method-to_list}{\code{Map$to_list()}}
\item \href{#method-clone}{\code{Map$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="is_field_allowed">}\href{../../stepfunctions/html/Block.html#method-is_field_allowed}{\code{stepfunctions::Block$is_field_allowed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="to_json">}\href{../../stepfunctions/html/Block.html#method-to_json}{\code{stepfunctions::Block$to_json()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="State" data-id=".next">}\href{../../stepfunctions/html/State.html#method-.next}{\code{stepfunctions::State$.next()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="State" data-id="accept">}\href{../../stepfunctions/html/State.html#method-accept}{\code{stepfunctions::State$accept()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="State" data-id="add_catch">}\href{../../stepfunctions/html/State.html#method-add_catch}{\code{stepfunctions::State$add_catch()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="State" data-id="add_retry">}\href{../../stepfunctions/html/State.html#method-add_retry}{\code{stepfunctions::State$add_retry()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="State" data-id="format">}\href{../../stepfunctions/html/State.html#method-format}{\code{stepfunctions::State$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="State" data-id="output">}\href{../../stepfunctions/html/State.html#method-output}{\code{stepfunctions::State$output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="State" data-id="update_parameters">}\href{../../stepfunctions/html/State.html#method-update_parameters}{\code{stepfunctions::State$update_parameters()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Map state class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Map$new(
  state_id,
  iterator,
  items_path = NULL,
  max_concurrency = NULL,
  comment = NULL,
  input_path = NULL,
  parameters = NULL,
  result_path = NULL,
  output_path = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state_id}}{(str): State name whose length **must be** less than or
equal to 128 unicode characters. State names **must be** unique
within the scope of the whole state machine.}

\item{\code{iterator}}{(State or Chain): State or chain to execute for each of the items in `items_path`.}

\item{\code{items_path}}{(str, optional): Path in the input for items to iterate over. (default: '$')}

\item{\code{max_concurrency}}{(int, optional): Maximum number of iterations to have
running at any given point in time. (default: 0)}

\item{\code{comment}}{(str, optional): Human-readable comment or description. (default: None)}

\item{\code{input_path}}{(str, optional): Path applied to the state’s raw input to
select some or all of it; that selection is used by the state. (default: '$')}

\item{\code{parameters}}{(dict, optional): The value of this field becomes the effective input for the state.}

\item{\code{result_path}}{(str, optional): Path specifying the raw input’s combination
with or replacement by the state’s result. (default: '$')}

\item{\code{output_path}}{(str, optional): Path applied to the state’s output after
the application of `result_path`, producing the effective output
which serves as the raw input for the next state. (default: '$')}

\item{\code{...}}{: Extra field names to pass to Block class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach_iterator"></a>}}
\if{latex}{\out{\hypertarget{method-attach_iterator}{}}}
\subsection{Method \code{attach_iterator()}}{
Attach `State` or `Chain` as iterator to the Map state, that
             will execute for each of the items in `items_path`. If an iterator
             was attached previously with the Map state, it will be replaced.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Map$attach_iterator(iterator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterator}}{(State or Chain): State or Chain to attach as iterator to the Map state.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allowed_fields"></a>}}
\if{latex}{\out{\hypertarget{method-allowed_fields}{}}}
\subsection{Method \code{allowed_fields()}}{
allowed extra fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Map$allowed_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert class to named list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Map$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Map$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
