% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_states.R
\name{Retry}
\alias{Retry}
\title{Retry base class}
\description{
A class for creating a Retry block
}
\section{Super class}{
\code{\link[stepfunctions:Block]{stepfunctions::Block}} -> \code{Retry}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Retry$new()}}
\item \href{#method-allowed_fields}{\code{Retry$allowed_fields()}}
\item \href{#method-clone}{\code{Retry$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="format">}\href{../../stepfunctions/html/Block.html#method-format}{\code{stepfunctions::Block$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="is_field_allowed">}\href{../../stepfunctions/html/Block.html#method-is_field_allowed}{\code{stepfunctions::Block$is_field_allowed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="to_json">}\href{../../stepfunctions/html/Block.html#method-to_json}{\code{stepfunctions::Block$to_json()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="stepfunctions" data-topic="Block" data-id="to_list">}\href{../../stepfunctions/html/Block.html#method-to_list}{\code{stepfunctions::Block$to_list()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Retry block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Retry$new(
  error_equals,
  interval_seconds = NULL,
  max_attempts = NULL,
  backoff_rate = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{error_equals}}{(list(str)): Non-empty list of strings, which match
`Error Names <https://states-language.net/spec.html#error-names>`_.
When a state reports an error, the interpreter scans through
the retries and, when the Error Name appears in the value of
a retrier’s `error_equals` field, implements the retry policy
described in that retrier.}

\item{\code{interval_seconds}}{(int, optional): Positive integer representing the
number of seconds before the first retry attempt. (default: 1)}

\item{\code{max_attempts}}{(int, optional): Non-negative integer representing the
maximum number of retry attempts. (default: 3)}

\item{\code{backoff_rate}}{(float, optional): A number which is the multiplier that
increases the retry interval on each attempt. (default: 2.0)}

\item{\code{...}}{: Extra field names to pass to Block class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allowed_fields"></a>}}
\if{latex}{\out{\hypertarget{method-allowed_fields}{}}}
\subsection{Method \code{allowed_fields()}}{
allowed extra fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Retry$allowed_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Retry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
