% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_choice_rule.R
\name{Rule}
\alias{Rule}
\title{Rule Class}
\description{
Class for creating a rule.
}
\section{Super class}{
\code{\link[stepfunctions:BaseRule]{stepfunctions::BaseRule}} -> \code{Rule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Rule$new()}}
\item \href{#method-to_list}{\code{Rule$to_list()}}
\item \href{#method-format}{\code{Rule$format()}}
\item \href{#method-clone}{\code{Rule$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Rule class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$new(variable, operator, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{(str): Path to the variable to compare.}

\item{\code{operator}}{(str): Comparison operator to be applied.}

\item{\code{value}}{(type depends on *operator*): Constant value or Path to compare `variable` against.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert class to list ready to be translated for
             Amazon States Language \url{https://states-language.net/spec.html}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
class formatting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
