% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepfunctions.R
\docType{package}
\name{stepfunctions-package}
\alias{stepfunctions}
\alias{stepfunctions-package}
\title{stepfunction: this is just a placeholder}
\description{
Step Functions Data Science `SDK` for building machine learning (`ML`) workflows and pipelines on `AWS`
             using the R `SDK` `paws`.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
